cCreate hook usePageInfo and clean up app.js
